Change to VS 2022. Involves updating to PlatformToolset v143 from v142. Rely on ABI compatibility for WxWidgets   (still using whatever version we had before), but we will be switching to 64 bit and DLL wxWidgets as part of this so hopefully a moot point. Tweak two source files to squash a couple of warnings. By the way the previous PGN2TDB commit was unnecessary, there were no source files in that directory, we just added unnecessary junk to the repo I think, sorry, will delete at some point.

+++ pgn2tdb/BinDb.cpp
@@ -1180,9 +1180,9 @@ static void sort_progress_probe()
             if( predicate_count>=predicate_nbr_expected || predicate_nbr_expected==0 )
                 permill = 1000;
             else if( predicate_nbr_expected > 1000000 )
-                permill = predicate_count / (predicate_nbr_expected/1000);
+                permill = (int) (predicate_count / (predicate_nbr_expected/1000));
             else
-                permill = (predicate_count*1000) / predicate_nbr_expected;
+                permill = (int) ((predicate_count*1000) / predicate_nbr_expected);
             predicate_pb->Permill( permill );
         }
     }
